Divide utils.test.js into smaller test files
